refactor: consolidate type utilities
