+++ src/components.tsx
@@ -3,8 +3,8 @@ import { Object3D } from "three"
 import { augment } from "./augment.ts"
 import { threeContext, useThree } from "./hooks.ts"
 import { manageProps, manageSceneGraph } from "./props.ts"
-import { type InstanceFromConstructor } from "./type-utils.ts"
 import type { Instance, Props, ThreeInstance } from "./types.ts"
+import { type InstanceFromConstructor } from "./types.ts"
 import { isInstance } from "./utils/is-instance.ts"
 import { withContext } from "./utils/with-context.ts"
