+++ src/types.ts
@@ -1,12 +1,6 @@
 import type { Accessor, JSX, Setter, Component as SolidComponent } from "solid-js"
 import type * as THREE from "three"
 import type { $S3C } from "./constants.ts"
-import type {
-  Constructor,
-  ConstructorOverloadParameters,
-  InstanceFromConstructor,
-  Overwrite,
-} from "./type-utils.ts"
 import type { Measure } from "./utils/use-measure.ts"
 
 interface ContextElements {
@@ -130,7 +124,7 @@ export type Instance<T = ThreeConstructors> = InstanceFromConstructor<T> & {
 }
 
 /** Metadata of a `solid-three` instance. */
-export type Metadata<T> = {
+export type Metadata<T extends object> = {
   props?: Props<InstanceFromConstructor<T>>
   children: Set<Instance>
 }
@@ -144,7 +138,7 @@ type MapToRepresentation<T> = {
 }
 
 /** Generic `solid-three` props of a given class. */
-export type Props<T> = Partial<
+export type Props<T extends object | unknown> = Partial<
   Overwrite<
     MapToRepresentation<InstanceFromConstructor<T>>,
     {
@@ -165,3 +159,88 @@ export type Props<T> = Partial<
     } & EventHandlers
   >
 >
+
+/**********************************************************************************/
+/*                                                                                */
+/*                                      Utils                                     */
+/*                                                                                */
+/**********************************************************************************/
+
+/** Generic constructor. Returns instance of given type. Defaults to any. */
+export type Constructor<T = any> = new (...args: any[]) => T
+
+/** Extracts the instance from a constructor. */
+export type InstanceFromConstructor<TConstructor> = TConstructor extends Constructor<infer TObject>
+  ? TObject
+  : TConstructor
+
+/** Omit function-properties from given type. */
+type OmitFunctionProperties<T> = { [K in keyof T]: T[K] extends Function ? never : K }[keyof T]
+/** Overwrites the properties in `T` with the properties from `O`. */
+export type Overwrite<T, O> = Omit<T, OmitFunctionProperties<O>> & O
+
+/**
+ * Extracts the parameters of all possible overloads of a given constructor.
+ *
+ * @example
+ * class Example {
+ *   constructor(a: string);
+ *   constructor(a: number, b: boolean);
+ *   constructor(a: any, b?: any) {
+ *     // Implementation
+ *   }
+ * }
+ *
+ * type ExampleParameters = ConstructorOverloadParameters<typeof Example>;
+ * // ExampleParameters will be equivalent to: [string] | [number, boolean]
+ */
+export type ConstructorOverloadParameters<T extends Constructor> = T extends {
+  new (...o: infer U): void
+  new (...o: infer U2): void
+  new (...o: infer U3): void
+  new (...o: infer U4): void
+  new (...o: infer U5): void
+  new (...o: infer U6): void
+  new (...o: infer U7): void
+}
+  ? U | U2 | U3 | U4 | U5 | U6 | U7
+  : T extends {
+      new (...o: infer U): void
+      new (...o: infer U2): void
+      new (...o: infer U3): void
+      new (...o: infer U4): void
+      new (...o: infer U5): void
+      new (...o: infer U6): void
+    }
+  ? U | U2 | U3 | U4 | U5 | U6
+  : T extends {
+      new (...o: infer U): void
+      new (...o: infer U2): void
+      new (...o: infer U3): void
+      new (...o: infer U4): void
+      new (...o: infer U5): void
+    }
+  ? U | U2 | U3 | U4 | U5
+  : T extends {
+      new (...o: infer U): void
+      new (...o: infer U2): void
+      new (...o: infer U3): void
+      new (...o: infer U4): void
+    }
+  ? U | U2 | U3 | U4
+  : T extends {
+      new (...o: infer U): void
+      new (...o: infer U2): void
+      new (...o: infer U3): void
+    }
+  ? U | U2 | U3
+  : T extends {
+      new (...o: infer U): void
+      new (...o: infer U2): void
+    }
+  ? U | U2
+  : T extends {
+      new (...o: infer U): void
+    }
+  ? U
+  : never
